Frontend-app/pyproject.toml
[tool.poetry]
name = "frontend-app"
version = "0.1.0"
description = ""
authors = ["Ole <105752308+Banquished@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.8"
django = "5.0.2"
docker = "7.0.0"
uvicorn = "0.27.1"
djangorestframework = "^3.15.0"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
